Adjustments load/unload
Changes to exitCheck.cs and sceneManager.cs.
Serialized field for exitCheck.cs refeer to the next scene to load and scene to unload when playeer passes through the exit. Therefore if player is currently going through exit  of level 2 towards level 3, next level to load should be  level_4 and level to unload should be level_1,
Persistent Elements scene now is set as the active scene where objects are instantiated. That is to say bullets and pulses will appear on this scene. Interactions with elements on another scenes still work.
Things to note, robots and robotHead/Wheel are tagged differently. If decided for one or the other check pulse and bullet scripts still work.
All scenes can be loaded while working on the unity environment without needing to unload any scenes before pressing play button.