Level exit teleport
Added end of level tp functionality on exitCheck.cs and playerTp.cs
Slight modifications to sceneManager.cs so we can have multiple scenes on the unity build and will unload everything but level 1 and persistent elements.
Changed PersistentElements name to Persistent_Elements for naming consistency.
Scripts are now placed on related files for better organization.
Updated GreenPowerUp prehab to be a trigger