First iteration of end screen with delay working
Moved the endsceern builder function outside of OnCollisionEnter2D in endTracker into a new function to be able to utilize the function to build the end game screen.
Currently using regular End Screen prefab but can be easily changed.
Added an "exit-door" empty gameobject to contain endTracker script in level 5 and use that to build final end game screen.